// src/data/types/base.ts
export interface Image {
  src: string;
  alt: string;
  width?: number;
  height?: number;
}

export interface Link {
  href: string;
  label: string;
  isExternal?: boolean;
}

export interface SocialLink extends Link {
  platform: 'linkedin' | 'twitter' | 'github' | 'dribbble' | 'behance' | string;
  icon: React.ReactNode; // Composant React pour l'icône
}

// Types pour les témoignages
export interface Testimonial {
  id: string;
  author: string;
  role: string;
  company?: string;
  content: string;
  rating?: number;
  image?: Image;
  date?: Date;
}